Allow passing `stim.PauliString` targets to `stim.circuit.append`
- `stim.circuit.append` now accepts `stim.PauliString` targets, which are auto-flattened into pauli targets and combiner targets
- Split the `circuit.pybind.cc` file into two files, so it takes less time to build when editing it
