+++ file_lists/pybind_files
@@ -1,4 +1,5 @@
 src/stim/circuit/circuit.pybind.cc
+src/stim/circuit/circuit2.pybind.cc
 src/stim/circuit/circuit_instruction.pybind.cc
 src/stim/circuit/circuit_repeat_block.pybind.cc
 src/stim/circuit/gate_target.pybind.cc
